backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.9
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1
openai==1.10.0
httpx
requests
python-dotenv
